Added in a few more functions with comments
These need to be added to the cli, or the cli needs to be abstracted
out into a more easy-to-use program so that all of this code isn't
being repeated. Ideally, Karma (#1) would be instituted and we could
verify each function as we go through it. There ought to be a visual
interface as well.